requirements.txt
django
python-decouple
psycopg2-binary
# gunicorn
django-allauth
# pytest
# pytest-django
# django-invitations
django-jazzmin
django-phonenumber-field[phonenumberslite]
django-import-export
tablib[xlsx,xls,csv]
django-ajax-datatable
PyJWT
cryptography
plotly
pandas